This update properly sets the exit code of the mysqlx when running in batch mode, the criteria is as follows:
JS/Python: return 1 if an error is found during the batch processing or 0 if no errors were found
SQL: Application stops processing at the first error found and return 1, if --force option is given it will continue no matter errors were found on SQL processing and the exit code will be 0.

Resolves: https://jira.oraclecorp.com/jira/browse/MYS-48

Additional Updates include:

Fix on the exception handlinh of SQL statement execution so they bubble up into JS context if the SQL was send from there.
Added Unit Test for the shell_core, tests the process_stream method used in batch processing.
Added test_utils.h with a class that properly saves the shell output (is a delegate implementation) so it can be used on checks in the different tests.
Updated some test cases to use the new utility class.

RB: 8812