Refactoring the resultset message reading. - Moving the logic to the X_resultset class. - Adding a parameter to the class constructor to identify whether the resultset should expect the last metadata message or not.
Context:
The Resultset class was designed in a way that the last message STMT_EXEC_OK containing some metadata is always received, as in SQL execution.
Other messages may produce a resultset and this message is not being sent, i.e. Crud Collection Find.

The message reading logic is meant to read the remaining data for a specific resultset, since now a resultset may differ depending on what created it, it is needed to "customize" how much data should be read from the transmission queue.