Updating packaging and application to properly handle when JS/PY are not available on build.
The application and build files have been updated so:
Start on JS mode if JS is available on build.
Start on PY mode if no JS is available on build but PY is available.
STart on SQL mode if no JS/PY are available on build.
Packages will not include the js modules if JS is not available.
When using -js/-py on startup and the mode is not available an proper error will be displayed.