MYS-435 - XSESSION HELP SAYS \SQL AVAILABLE WHEN IT IS NOT
Problem
While creating an X session, help displays SQL mode when it is not
available. When `\sql` command is used to switch to SQL mode in
an X session, the appropriate error message is not displayed.

Solution
Modified the `switch_shell_mode` method of `Interactive_shell` so that
while switching to SQL mode, correct session value is retrieved.
Modified the `command_loop` method of `Interactive_shell` to ensure
appropriate help message is displayed based on the session.