Fixing Cluster.status()/Cluster.describe()
These functions were returning a JSON string originally.

Later it was decided to return a JSON object, to provide a quick solution we decided to hack the original code and parse the produced JSON string to generate the JSON object (A Dictionary).

This refactor is to generate the JSON object directly rather than doing the JSON Str->JSON Object translation, at the same time it fixes the next issues:
- MYS-708: status now reports the replicaset name
- MYS-850, 926 and 945: Related to the wrong nesting of the RO instances
- Removed the adminType component on describe()

RB: 14477