Fix issues related to addInstance on localhost
Bug #25027264
CAN'T CONNECT TO X PORT

Bug #25047488
SHELL RESOLVES SERVER HOSTNAME AS LOCALHOST INSTEAD OF
MACHINE NAME/IP

Bug #24832550
ADDINSTANCE() FAILS FOR REMOTE INSTANCE

When addInstance() is executed, ensure that metadata contains the
public hostname of the instance instead of localhost, specially when
the shell is connected to localhost. This is done by querying the
performance_schema tables from the GR plugin, so we get the same
hostname used by GR.

Because often only root@localhost exists, attempting to connect to a
MySQL server using the public hostname will result in access denied.
To work around that, we detect when the seed instance to be used is
actually running on localhost, then ensure that localhost is used in
place of the public hostname

Added a allowRootFrom option to deploySandboxInstance(). In interactive,
if the option is not provided, the user is queried for it.