BUG#25077429: The uri should also be a positional argument
This patch changes the way mysqlsh executable handles the positional
argument. Previously mysqlsh executable handled the positional
argument as being a the default schema name to connect to.

This patch changes that behavior, now the positional argument is
parsed as being an URI to connect to. The --uri option however remains
unchanged and can still be used.