BUG#25056802: portx option is not used for startlocalinstance
The help of several SandboxInstance operations (start, stop, delete
and kill) was incorrect. It displayed options that should not be
presented. Furthermore for those functions the shell accepted
parameters such as portx, even if they were never used on the backend
when mysqlprovision is called.

This patch removes the help text of incorrect options and fixes the
argument checking of the SandboxInstance functions so that they throw
an error in case any unused argument is provided by the user in the
options dictionary.