BUG#24380669 INCORRECT SINGLE QUOTE ESCAPE HANDLING UNDER CLASSIC SQL MODE
Problem
When more than two consecutive single quotes or double quotes are included
in commands in classic sql mode, the behavior in the Shell is not as expected.

Solution
In `determineStatementRanges` method in `utils_mysql_parsing.cc`, the code
is modified to handle consecutive single and double quotes. Also modified
the unit tests to include such cases.

RB: 14697