BUG#25254905 dba.stopsandboxinstance command does not work with mp sandbox stop BUG#25091818 : sandbox stop fails for windows zip package
For windows systems sending SIGTERM signal to the server doesn't allow
for the server to have a correct shutdown like it does on unix.  This
patch removes the use of the SIGTERM command for the sandbox stop
command, replacing it with the use of the SQL command SHUTDOWN
introduced in MySQL 5.7.9.

Furthermore, this patch adds support for the AdminAPI stop command to prompt for
the root password (if not provided on the options dict) and pass it to
the mysqlprovision executable.