BUG#25226225: non ssl-enabled server cannot join cluster even if gr ssl not requested BUG#25306104: replace --skip-ssl by --ssl-mode
This commit fixes two issues since BUG#25306104 was dependent on the fix for BUG#25226225.
Previously non SSL enabled servers were unable to join a Cluster where GR ssl support
is disabled altought servers on that cluster support SSL.
This patch fixes this incorrect behavior and also replaces the --skip-ssl option used on
the start and join operations for a new option --ssl-mode.