BUG#25223255: RECOVERY USER AND REQUIRE SSL
There was an inconsistency between the way the replication users were created
by mysqlprovision (with REQUIRED SSL) and the X Shell (without REQUIRED SSL)
when SSL was enabled for Group Replication.

This patch fixes that issue by removing the use of REQUIRED SSL to create
the replication users using mysqlprovision, independently if SSL is enabled
or not to configure Group Replication. The use of 'REQUIRED SSL' is not
mandatory and secure connections are still able to be performed (using SSL)
even if the user was created without REQUIRED SSL.