BUG#25298569 add an option to specify the ip whitelist for the cluster BUG#25298445 add an option for join and start replicaset to specify the ip whitelist for gr
This patch adds to ngshell and to mysqlprovision a new ipWhitelist option
that allows users to set the group_replication_ip_whitelist variable used
by Group Replication.

This is important to allow users to use Group Replication (GR) on non-private
networks (e.g., in the cloud). By default, if not specified explicitly, the
whitelist is automatically set by GR to the private network addresses that
the server has network interfaces on.
This is also important for security reasons, allowing the user to restrict
the IP addresses (hosts) accepted by GR.