BUG#25064894: SSL SUPPORT REQUIRED FOR CLUSTER DEPLOYMENT
This follow up patch adds a new memberSslMode option to configure SSL
(encryption) for the cluster. This option will be used to set the appropriate
SSL mode for mysqlprovision that will configure Group Replication for the
target instance.

The new option will be available for createCluster(), addInstance() and
rejoinInstance() and it will support the following values: AUTO, REQUIRED,
and DISABLED. By default, AUTO will be used for the SSL Mode and SSL will
be automatically enabled/disabled based on the SSL support for the instance
and the SSL settings already used for the cluster.

The previous memberSsl* options to setup SSL for the cluster (Group
Replication) were removed.

Existing unit tests were updated accordingly.