BUG#24594156: NO HA, METADATA NOT REPLICATED
The AdminAPI functions to create a cluster and add instances to it
were not verifying if a binlog filter (binlog-do-db or binlog-ignore-db) was
defined by the user that would prevent the metadata database
(mysql_innodb_cluster_metadata) from being replicated for all member of the
cluster.

The metadata must be replicated to all members of the cluster and no
binlog filters that prevents it are allowed.

This patch fixes that issue by adding a validation, for the functions to
create the cluster and add instances, that verifies if any invalid
binlog filter (binlog-do-db or binlog-ignore-db) is defined and issue an
appropriate error in that case.

New unit tests were added to verify this issue.