BUG#25096853: DEPLOYSANDBOXINSTANCE() REMOVES BACKSLASHES IN MESSAGES
The dba.deploySandboxInstance() function when executed in verbose mode,
removed backslashes (\) for the output messages resulting in strange messages.
This issue is observed because backslash generated by the mysqlprovision
output in JSON are not correctly escaped, using a double backslash (\\).

This patch fixes that issue by handling some special control characters
correctly in the JSON output produced by mysqlprovision. More specifically
'\' are converted to '\\' and '"' are converted to '\"'.

The existing logger unit test for mysqlprovision was update to cover this
issue.