BUG#25469458: USE MYSQLD DEFAULT PATH FOR DEPLOYSANDBOXINSTANCE()
Currently, the dba.deploySandboxInstance() function expects the path to the
mysqld binary to be set in the PATH environment variable. This might be
inconvenient for inexperienced users that use default settings, and it would
be expected for this function to also try to search for mysqld in the default
MySQL installation path.

This patch improves the behavior of the dba.deploySandboxInstance() when
searching for a valid mysqld file to use, now also considering the default
MySQL installation path beside the PATH environment variable.

In particular, when search for the location of 'mysqld' the
dba.deploySandboxInstance() function now also looks for the executable in the
following locations:
- Windows:
  - C:/Program Files/MySQL/MySQL Server 5.7/bin
  - C:/Program Files/MySQL/MySQL Server 8.0/bin
- POSIX (Non-Windows):
  - /usr/sbin
  - /usr/local/mysql/bin/
  - /usr/bin/
  - /usr/local/bin/
  - /usr/local/sbin/
  - /opt/local/bin/
  - /opt/local/sbin/

Note: additional location added based on the information found in the MySQL
reference manual for default location used for the mysqld file by other tools
(e.g., mysql_install_db).

In addition, this patch also fixes an issue when the mysql_ssl_rsa_setup
tool was not found (even using the option to ignore SSL errors), raising an
incorrect error.

The existing unit tests for mysqlprovision were updated accordingly.