BUG#25096853: DEPLOYSANDBOXINSTANCE() REMOVES BACKSLASHES IN MESSAGES
This follow-up patch fixes an issue in the custom logger JsonFormatter, that
was not handling the special case of a dictionary being specified as argument
of the log message. In this case, the arguments parser of the message is
skipped. Dictionaries should not be used as arguments for log message since
they internally convert \ to \\ and can lead to duplicated backslashes.

The existing logger unit test for mysqlprovision was update to cover this
issue.