Help Engine: Added support for the <TAG>_PARENTS construct on the help generation engine to properly geerate help() on inherited members.
In the past, any documented member in a base class was not being shown on the help() output for a descendant class.

With the _PARENTS construct a new way of resolving help is introduced, by searching from bottom up in the class hierarchy until information is found for a member.

Fixes Bug #24964623 - DB.GETSCHEMA() DOES NOT WORK (FROM A USER PERSPECTIVE)
- Not having this caused the help on the NodeSession being empty, causing the confussion described on the issue