BUG#24756046: MYSQL SHELL IS NOT IN SYNC WITH GR FOR MULTIMASTER SETUP
The MySQL Shell was not showing the status of a cluster correctly, if the
topology type was changed manually (for example, from primary master to
multi-master). This was due to the fact that the cluster.status() function
was relying on the topology type in the metadata to display the cluster
information without verifying the real topology being used.

This patch fixes this issue by checking the current topology type of the
cluster (same for all cluster instances) before checking the cluster status
and issuing an error if it does not match the topology type expected in the
metadata.

New unit tests were added to verify this issue.