BUG#25141048: cluster object is not invalidated after dissolve()
With this patch an informative error message is raised when
trying to access a function of member of a dissolved cluster.

This patch also fixes a small issue with the shell_script_tester
that could lead to a segmentation fault due to a out of bounds
access of an array.