BUG#24598272: DBA.DEPLOYLOCALINSTANCE() CAN'T START ERROR HIDES TRUE CAUSE
The error message for the dba.deploySandboxInstance() when a valid mysqld
is not available, only indicates that mysqld was not found without providing
additional information about the possible cause (e.g., version of available
mysqld is not supported).

This patch improves the error message issued by the deploySandboxInstance()
function by providing additional information to the user about the possible
cause of the issue. A different error is issued if no mysqld at all is found,
or if no valid mysqld was found (despite some invalid mysqld being located).

The unit tests for mysqlprovision were updated accordingly.