BUG #24594075: dba.createcluster / dba.validateinstance: setting server variables
In order to make GR able to run, createCluster() and/or addInstance()
may need to change server variables. The AdminAPI needs to have the
ownership of the instances to ensure the setups work and at the same
time hide the complexity of it from the user. However, those changes
cannot be hidden from the user.
If the defined workflow of using dba.checkInstanceConfiguration() on the
instances before any operation is not followed, the AdminAPI must print
a warning on the interactive mode stating that the configuration must be
changes in order to create an InnoDB Cluster and in top of that ask for
confirmation to proceed the operation. On the other side, all of the
changes that happen must be logged.

This patch fixes the problem in several levels:

  - MP extended to output logging messages of Warning level indicating
    each variable change
  - dba.createcluster() and Cluster.addInstance():
    - Catch MP log and log it on the Shell side
    - Added internal call for checkInstanceConfiguration() on the
      interactive level to verify if changes need to be done, warn the
      user and ask for confirmation
  - Added log checking mechanism to be used on the unit-tests:
    - Implemented log_hook on the unit-tests base class to append log
      messages of the log-level set on an internal structure
    - Added log validation functions to be used with the macros:
      MY_EXPECT_LOG_CONTAINS(x) and MY_EXPECT_NOT_LOG_CONTAINS(x)
  - Added unit-tests to validate the following scenarios:
    - Non-Interactive mode: confirm that changes are logged in
      misconfiguration scenario and opposite
    - Interactive mode: confirm that prompt is shown and changes are
      logged in misconfiguration scenario and opposite