BUG#25789094: sandbox creation does unnecessary operations and misuses file descriptors
There is a problem caused by the value of the close_fds parameter (by
default False on python2.7) which on windows causes all files opened
by the shell to be inherited by the mysqld executable.  As a result,
one cannot remove those files. This causes issues with the installer
and temporary files savepoint.
This patch fixes this issue and also allows changes the create operation
to allow the server to be left running instead of shutting it down.