BUG#24580777: DEPLOY SANDBOX INSTANCE IN FOLDERNAME WITH SPACES DOES NOTHING
The dba.deploySandboxInstance() function does not work as expected when
a directory containing spaces is specified for the 'sandboxDir' parameter,
for example: "/home/sandbox/foo bar".

This patches fixes this by quoting the path value properly at the
mysqlprovision level in order to use the correct value for the sub processes
execution.

A new unittest was added to cover this issue.