BUG 24977542: CHECKINSTANCESTATE() SIGNATURE IS DIFFERENT IN HELP THAN THE ACTUAL ONE
Problem:
While checking the parameters of checkInstanceState(), the number of
parameters were expected to be 0 or 1. This resulted in inconsistent
behavior as number of parameters in checkInstanceState() should be 1
or 2. No parameter should throw an error, while the second optional
parameter is the password for the instance.

Solution:
While checking the parameters of checkInstanceState(), the number of
parameters are now expected to be 1 or 2.