BUG#25975317: PYTHON VERSION ERROR IS NOT PROPERLY EXPOSED
When a AdminAPI command is used with a Python version that does not meet
the requirements (> 2.7 and < 4), a RuntimeError is issued but with an empty
message.

This patches fixes this problem by issuing the error message (from
mysqlprovision) in a recognized JSON format so that it can be properly
displayed to the users.