BUG#25932403: CHECKINSTANCECONFIGURATION IS NOT (ALWAYS) RELIABLE
The dba.checkInstanceConfiguration() command is failing to report the
needed requirements to use Group Replication when Multi-Threaded Slave (MTS)
is enabled for the server instance. Consequently, dba.createCluster() fails
because Group Replication does not start.

If MTS is enabled, i.e. slave_parallel_workers > 0, then the following
settings are required for Group Replication to work (other settings are not
currently supported):
- slave_parallel_type=LOGICAL_CLOCK
- slave_preserve_commit_order=ON

This patch fixes this issues by adding the needed support to verify the
requirements for Group Replication when MTS is enabled, reporting it
properly for dba.checkInstanceConfiguration() and updating the settings
dynamically when required (e.g., dba.configureLocalInstance(),
dba.createCluster(), cluster.addinstance()).

New unit tests were added to verify this issue (MTS requirements for GR).