BUG #25223298: WRONG DONOR USED AND ERROR MSG WRONG
On a scenario on which a new primary instance is elected, adding a new
instance to a cluster resulted on an error due to a failed connection to
the previous primary instance. This was caused by an erroneous and
incomplete implementation of the function to obtain the seed instance of
a replicaset.

This patch fixes that function in order to:

  - If single-primary mode is enabled the seed is obtained from
    "group_replication_primary_member"
  - If multi-primary mode is enabled the seed is the first instance
    obtained from the list of online instances

Added unit-tests.