BUG#25614855 : CONFIGURELOCALINSTANCE URI USER WITHOUT PERMISSIONS, CREATES A WRONG NEW USER
The configureLocalInstance command had an issue where it didn't check if the user
account being used to create a new user for managing a GR group had the
required permissions to do so. This patch fixes this issue by checking for those permissions
when the shell user decides to create a new MySQL user.