BUG#25485035: TEST FAILS ON SANDBOXDIR PATH TOO LONG
The deployment of sandboxes stops with an error if a sandbox directory with
a long path is specified. This is due to a server limitation for the
specification of the socket path. A specific socket value is required to
avoid different sandbox instances from creating the socket file on the same
(default) location.

This issue was fixed by only setting the socket file name (i.e., mysqld.sock
and mysqlx.sock) instead of the full absolute path to the socket file in the
sandbox option file. When no absolute path is specified the server creates
the socket file in the data directory (which is different for each sandbox
instance). The existing sandboxDir validation was removed.

A new unit tests was added to cover this issue.