BUG #25404009: ERROR ON REMOVE INSTANCE LEADS TO LISTING STOPPED, BUT STILL REPLICATING
If removeInstance() fails due to a connection error of some kind, the
error was reported but the instance was removed from the Metadata and
kept on the GR group. That was caused due to the metadata update
happening before the GR group update. The instance was removed from the
metadata before any replication update, which could fail for some reason
leaving the Metadata inconsistent.

This patch fixes the bug by making sure the metadata is updated
accordingly depending on the GR update status.

Added unit-tests