BUG #25534693: RESCAN DETECTS AN INSTANCE ALREADY PART OF THE CLUSTER, RELATED TO LABEL
This patch fixes a random behaviour of Cluster.rescan() which may
produce a wrong JSON output by having an instance already part of the
cluster belonging to the newlyDiscoveredInstances list and to the list
unavailableInstances.
The interactive layer would also behave erroneously by prompting the
user if wants to add/remove the instance from the cluster.

Added unit-tests