BUG#25226130 : REMOVAL OF SEED NODE BREAKS DISSOLVE: LOGICERROR RETURNED
The removal of the last instance of a cluster using the removeInstance()
command leaves the cluster in an inconsistent state, not allowing other
operation to be performed (like dissolve). This happens because the
remotion of the last instance should dissolve the cluster (also removing
the metadata).

This patches fixes this issue but not allowing the removal of the last
instance of the cluster using removeInstance(). If removeInstance() is
used an error is issued indicating that dissolve() should be used instead
to remove the last instance and dissolve the cluster.

New unit tests were added to verify this issue (for regression).