BUG#25516390: REMOVE INSTANCES LIST DOESN'T VALIDATE THE LIST,PROMPTS USER FOR REJOIN/REMOVE
This patch fixes an issue with the interactive layer of the
rebootClusterFromCompleteOutage command.

The interactive layer, when prompting for the instances to rejoin the
cluster, was not ignoring the instances already specified using the
removeInstances list. This made it possible for the same instance to
be specified twice, once in the removeInstances list (manually
specified by the user) and once in the rejoinInstances list
(automatically added by the wizard of the interactive layer).

With this patch, instances that are eligible to be added to the
rejoinInstances list but that are specified in the removeInstances
list are skipped by the wizard.

This patch also adds validation on the non interactive layer for the
rejoinInstances list that throws an error if unreachable instances are
specified.