BUG#25519190 : CONFIGURELOCALINSTANCE() FAILS UNGRACEFUL IF CALLED TWICE
The dba.configureLocalInstance() command fails if called twice, using the
same settings, because the second time it is called it tries to re-create
a user that already exists. If all is already setup properly (e.g., user
already available with the proper privileges) then no error should be issued.
In addition, the user creation error was not reported in the JSON returned
output.

This patch fixes this issue by validating if the cluster administration user
already exists with the appropriate privileges, and reporting an user
creation error in the JSON returned object.

New unit tests were added to verify this issue (for regression).