BUG #25786495: CLUSTER.REJOININSTANCE() DOESN'T USE THE USER IN THE CONNECTION STRING PARAMETER
rejoinInstance() needs a connection to the seed instance for
validations and to use the instance as peer for the rejoining instance.
The current implementation of rejoinInstance() assumes that a
cluster was created using the 'root' account, so the connection to the seed
instance assumes the 'root' user. If the user is using a different user
than 'root', the rejoinInstance() command will fail.

This patch fixes the issue and allows the use of a different account
than 'root'.

Added unit-tests