BUG#25966057: SQL_MODE ANSI_QUOTES BREAKS DBA.CREATECLUSTER() COMMAND
If the ANSI_QUOTES SQL mode is enabled, string literals can be quoted
only within single quotation marks because a string quoted within
double quotation marks is interpreted as an identifier. This was the origin
of this bug and this patch replaces all occurrences of double quotes with single
quotes on string literals used by the shell.