Fixed regression in resultset printing
Resultsets with multiple columns got the value of the first column
printed for all columns. Regression from zerofill work.

Also fixed zerofill handling to not store 2 copies of every field value
just for the case of needing to handle zerofill, which was moved to the
display code.

Added a zerofill property for Column metadata.