BUG#24916064 : CAN NOT REMOVE STOPPED SERVER FROM A CLUSTER
The operation removeInstance() does not work on stopped instances and it is
not possible to remove an non available instance from the cluster.

This patch fixes the above issue by improving the error message showed when
an instance is not available or not reachable (e.g., MySQL server stopped).
In addition, a new option 'force' was added to the removeInstance() command
to allow users to remover instance (from the metadata) that are permanently
not available, avoid obsolete data from being kept in the metadata of the
cluster. The help of the removeInstance() was also updated accordingly.

New unit tests were added for regression and to exercise the new features.