BUG #25724564 - SHELL GENERATES RANDOM SERVER-ID, WHICH GR REUSES FOR AUTO_INCREMENT_OFFSET
This patch implements the following logic for setting
auto_increment parameters:

IF single-primary-mode
   set auto_inc_inc = 1
   set auto_inc_offs = 2
ELSE
   set auto_inc_inc = 7
   set auto_inc_offs = 1 + server_id%7
ENDIF

This is done when creating a cluster and joining an instance to it.
configureLocalInstance() will also persist these sysvars.

Problem
-------
Original bug claimed that auto_increment would be prone to overflow due
to it being set to a very high default value by GR, but that is
incorrect.

In reality, setting auto_increment_offset to a value higher than
auto_increment_increment will cap the value to a relatively low value
with litle risk of overflow. The capping however happens in strange
ways, which is confusing and thus, it is preferrable (but not strictly
mandatory) that the offset is always set to a value lower than
auto_increment_increment.

Duplicate auto_increment_offset values are still possible in some cases,
but not an issue. They will increase the probability of a collision with
concurrent transactions happening at different members, but that is a
known limitation of GR.

This patch works around the following issues in the GR plugin:

- auto_increment_increment is set to 7, even in single-primary mode,
  where the problem of colliding auto_increment values does not exist
- auto_increment_offset is set to server_id, even if its higher than
  auto_increment_increment