BUG#25966731 : ALLOW-NON-COMPATIBLE-TABLES OPTION DOES NOT EXIST
When adding nodes to a InnoDB Cluster using the appropriate AdminAPI
operations, some checks are performed to verify the compatibility of the
existing tables. If incompatible tables (e.g., MyISAM tables) are detected
then an error is issued, however a message is logged referring an option
not available for the AdminAPI operations: --allow-non-compatible-tables.
This option is from mysqlprovision (internal tool) and should not be
displayed to users.

This patch fix this issue by removing the logged message including the option
from mysqlprovision.

New unit tests were added for regression.