BUG#25226117 : GR SSL status cannot be queried
There is no AdminAPI command that allows to verify the SSL mode being used by
the cluster (defined with the memberSslMode parameter).

This patch adds a new parameter 'ssl' to the JSON document returned by
the AdminAPI <cluster>.status() command with the information of the SSL mode
being used by the cluster (replicaset). The values returned by this parameter
correspond to the value of the 'group_replication_ssl_mode' server variable.
Currently, only the following values are supported by the AdminAPI, being
configured through the memberSslMode for the operations to create the cluster
and add instances to it: "REQUIRED" and "DISABLED".

Existing unit tests were changed to include the new SSL mode information
returned by the status() command.