BUG#25495056 : ERROR MESSAGE WHEN GR SSL MODES MISMATCH IS MISLEADING
There are a few small typos in the error messages issued when the SSL mode
used by the cluster and the one specified by the user for the addInstance()
command mismatch. This patch fixes those typos.

Existing unit tests were updated to check the text of the corrected error
messages.