BUG#25526248: DBA.CONFIGURELOCALINSTANCE() CAUSING SECURITY ISSUES, WORSENING ACCESS PERMS
This patch changes the permissions with which the backup files of
modified configuration files are created.
Now, the backup configuration file's permissions are strictest between 640 and
the permissions of the original configuration file. In other words, with this
patch backup_file_perms = (original_file_perms & 640), e.g. if the
original file had permissions 700, the backup file will have permissions 600.