WL#11054 - AdminAPI: super-read-only handling
Starting with MySQL Server 8.0.2, whenever the GR plugin is stopped
super_read_only is enabled (WL#10611) to prevent any client to execute
write transactions in the instance. Due to this behaviour change, it's
not possible to add or rejoin an instance to a Cluster which previously
belonged to a GR group (or InnoDB Cluster).

This patch implements WL#11054 which extends the following AdminAPI
commands with a new option 'clearReadOnly' that allows disabling
super_read_only:

  - dba.configureLocalInstance()
  - dba.createCluster()
  - dba.rebootClusterFromCompleteOutage()
  - dba.dropMetadataSchema()

The interactive layer is also extended in order to provide the best
usability by warning and prompting the user accordingly.

BUG#26422638 is fixed by removing the unnecessary creation of the
replication-user on the joining instance and leave it for the seed
instance.

Updated existing tests and added new tests to cover the added
functionalities.

Closes:
    - BUG#26422638 - CAN'T ADD/REJOIN INSTANCES WHICH WERE PART OF A
CLUSTER DUE TO SUPER-READ-ONLY