Fixed rejoinInstance() issues regarding the seed instance:
rejoinInstance() was assuming that the current session is the
session to the seed instance when that may not be true since the
function can be executed with a session to a non seed instance
(R/O).

The fix is to use the get_peer_instance() which obtains the seed
instance for both single-primary and multi-primary modes.

Also added missing call for wait_slave_state() on AdminAPI python
test and did code cleap-up of the rejoinInstance() function.