BUG#26818744 : MYSQL SHELL DOES NOT ADD THE SERVER_ID ANYMORE
Starting from MySQL server 8.0.3 the binary logging is enabled by default and
consequently the server ID default value also changed to 1 (instead of zero).
The server ID verification performed by the AdminAPI is not considering this
change, only considering server IDs with the zero value invalid.

This can lead to issues since the server ID must be unique for of all servers
in a cluster. Therefore, if the default server ID is still used, it should be
considered invalid, requiring the user to change it or allow
dba.configureLocalInstance() to change it properly.

This patch fixes this issue by doing a different server ID verification for
MySQL server versions > 8.0.3. More specifically, it is verified if the
server ID value is still the server default at compile time and a proper
error is issued by in that case (like for when the server ID is zero), so
that the dba.configureLocalInstance() can set that value with a random one
in that case.

System test added for regression.