BUG#26834542 : WRONG RECOVERING STATE CAUSES ADDINSTANCE TO FAIL
Sometimes (not deterministic, depending from the server version and recovery
time) the addInstance() function fails with an error indicating that the
peer server is in RECOVERING state despite being ONLINE. The check of the
peer (seed) instance state is incorrect, sometimes the state of another
instance is checked instead of the one of the seed.

This patch fixes this issue by correcting the function (query) used to obtain
the instance state, making sure that only the state of the target instance
is returned.

No new test were added because existing test already cover this issue. This
issue is likely to be the cause of some none-deterministic failures for the
existing tests.