BUG#25676405: PATH EXPANSION IS NOT SUPPORTED ON THE SHELL
Add support for tilde expansion. If path begins with `~`, all of the
characters up to first OS specific separator are considered as
a tilde-prefix. All characters after `~` are considered as `loginname`.

If `loginname` is empty then:
  On Unix: `HOME` environment variable is returned if is set, otherwise
  current user home directory is looked up in password directory.

  On Windows: `HOME` or `USERPROFILE` or `HOMEDRIVE`+`HOMEPATH`
  environment variable is returned if is set, otherwise user home
  directory is obtained using `FOLDERID_Profile` `KNOWNFOLDERID`.
  If all of above fail, empty string is returned.

If `loginname` is not empty then:
  On Unix: User home directory path associated with `loginname` is
  looked up directly in the password directory.

  On Windows: Retrieving home directory of another user on Windows
  platform is *NOT SUPPORTED*.

If the expansion fails or if the path does not begin with `~`, the
path is returned unchanged.