BUG #26485316: CREATING AN INNODB CLUSTER FROM AN EXISTING GROUP REPLICATION NOT WORKING
Creating an InnoDB Cluster from an existing GR group, by using the
adoptFromGR option in the dba.createCluster() commandi, would fail with an
error stating that the instance is already part of a replication group.
The issue was only present in the Shell default wizard mode.

This patch fixes an issue in interactive layer of the dba.createCluster()
command, allowing the use of adoptFromGR.

Added new tests to cover the usage of adoptFromGR in both interactive
and non-interactive modes.