BUG#26609909: MYSQL-SHELL DBA.CHECKINSTANCECONFIGURATION()
This patch fixes issues with the checkInstanceConfiguration and
configureLocalInstance commands.

No check was done to confirm if the account used to run the command
had enough privileges to actually execute it. This issue was present
in both the checkInstanceConfiguration and configurelocalinstance
commands. With this patch, the first operation done in any of the
commands (besides input validation) is checking if the provided
account has the required privileges and a runtime error is thrown in
case it doesn't.  Furthermore, a new privilege, the SELECT privilege
on all schemas was added to the list of privileges to check.

Also, the help options presented in checkInstanceConfiguration were
incorrect and showed extra options that the command did not accept.

Finally, the SELECT privilege on all schemas was also added to the
list of privileges granted to the clusterAdmin account created using
the configureLocalinstance command.

This patch requires the list of privileges required by the
clusterAdmin account to be updated on the manual
(https://dev.mysql.com/doc/refman/en/mysql-innodb-cluster-working-with-production-deployment.html),
replacing the "GRANT SELECT ON sys.* TO your_user@'%' WITH GRANT OPTION;"
" with "GRANT SELECT ON *.* TO your_user@'%' WITH GRANT OPTION;"