BUG#26830224: DBA.CONFIGURELOCALINSTANCE() CLAIMS STATUS IS OK WHEN MY.CNF UPDATE FAILED
When using the configureLocalInstance command to update the
configuration file of a server already belonging to a cluster no error
message would be displayed if the user didn't have enough file system
privileges to update the configuration file.
This patch fixes that issue and now, an appropriate error message is shown
in the output if the user is lacking privileges to update the configuration
file.
Furthermore, some test related code was moved.
A new test was added to verify the new behavior, although it only runs on
Unix systems.