BUG#26248116 : MYSQLPROVISION DOES NOT USE SECURE CONNECTIONS BY DEFAULT
The mysqlprovision internal component of the AdminAPI was not supporting
secure connection by default due to a known limitation from connector-python.
However, starting from connector-python 2.1.7 secure connections are
established by default and that version (or higher) should be used for
mysqlprovision.

This patch add unit tests for regression to ensure that a version of
connector-python with support for secure connection by default is used.