WL#11162 : Add options to set GR address, seeds and name for a cluster
This patch adds new options for the createCluster() and addInstance()
functions to allow advanced users to specify the group name, local address,
and group seeds settings for Group Replication. The value for those
options will override the ones automatically determined by the AdminAPI when
not specified.

In more detail the following options were added to the createCluster() and
addInstance() functions:
- 'localAddress' used to set the 'group_replication_local_address' variable;
- 'groupSeeds' used to set the 'group_replication_group_seeds' variable:
- 'groupName' only available for createCluster() and used to set the
  'group_replication_group_name' variable;

This patch also implements the feature requested by BUG#26485254 and
BUG#26838005.

Tests were added to ensure the defined requirements are meet and for
regression.