BUG#26870329: CLUSTER.REJOININSTANCE() SHOULD ENSURE MEMBER IS NOT ALREADY IN GROUP
The rejoinInstance operation will attempt to rejoin an instance even
if the member being rejoined is already part of the group.

This patch adds a new check that validates if the instance being added
is missing from the group, in which case it is rejoined. Otherwise a
runtime error is thrown.
Interactive documentation was updated.

New tests were added and existing tests were updated.